changed floats to ints

+++ testDepth.py
@@ -116,7 +116,7 @@ try:
         if not depth:
             continue
 
-        objectCenter = (bbox[0]+bbox[2]/2, bbox[1]+bbox[3]/2)
+        objectCenter = (int(bbox[0]+bbox[2]/2), int(bbox[1]+bbox[3]/2))
         centerDepth = depth.get_distance(objectCenter[0], objectCenter[1])
 
         speed = 6000
